add repo mongo ping implementation
